refactor: Computer 클래스 삭제, getComputerNumberList 함수 BaseballGame 클래스로 이동
